+++ objc2-encode/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+* Added `Encoding::LONG` and `Encoding::U_LONG` to help with platform
+  compatibility; use these instead of `c_long::ENCODING` and
+  `c_ulong::ENCODING`.
+
 
 ## 2.0.0-beta.2 - 2022-01-03